backend/new-architecture/ml-service/requirements.txt
# Shine Skincare App - ML Service Requirements
# TensorFlow and image processing dependencies

# Core ML framework
tensorflow==2.13.0

# Image processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Scientific computing
numpy==1.24.3
scipy==1.11.1

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0

# WSGI server for production
gunicorn==21.2.0

# Environment variable management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.1.0

# Health check utilities
healthcheck==1.3.0

# Production server
waitress==2.1.2

# Additional ML utilities
scikit-learn==1.3.0
